# swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/users": {
      "post": {
        "operationId": "UsersController_create",
        "summary": "Create a new user",
        "description": "Creates a new user account. Only accessible by administrators.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateUserDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "User created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "409": {
            "description": "User with email or username already exists"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "get": {
        "operationId": "UsersController_findAll",
        "summary": "Get all users",
        "description": "Retrieves a paginated list of users with optional filtering by role and status.",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (default: 1)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (default: 10)",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "role",
            "required": false,
            "in": "query",
            "description": "Filter by user role",
            "schema": {
              "enum": [
                "ADMIN",
                "MANAGER",
                "USER"
              ],
              "type": "string"
            }
          },
          {
            "name": "isActive",
            "required": false,
            "in": "query",
            "description": "Filter by active status",
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Users retrieved successfully"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/users/me": {
      "get": {
        "operationId": "UsersController_getCurrentUser",
        "summary": "Get current user profile",
        "description": "Retrieves the profile of the currently authenticated user.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_updateCurrentUser",
        "summary": "Update current user profile",
        "description": "Updates the profile of the currently authenticated user. Users can update their username, first name, last name, and avatar.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "409": {
            "description": "Email or username already exists"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/users/statistics": {
      "get": {
        "operationId": "UsersController_getStatistics",
        "summary": "Get user statistics",
        "description": "Retrieves comprehensive user statistics including totals and role distribution.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Statistics retrieved successfully"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/users/{id}": {
      "get": {
        "operationId": "UsersController_findOne",
        "summary": "Get user by ID",
        "description": "Retrieves a specific user by their unique identifier.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User unique identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User found successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "patch": {
        "operationId": "UsersController_update",
        "summary": "Update user by ID",
        "description": "Updates a specific user account. Full access for admins.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User unique identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "User updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UserResponseDto"
                }
              }
            }
          },
          "404": {
            "description": "User not found"
          },
          "409": {
            "description": "Email or username already exists"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      },
      "delete": {
        "operationId": "UsersController_remove",
        "summary": "Deactivate user",
        "description": "Deactivates a user account (soft delete). Only accessible by administrators.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User unique identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User deactivated successfully"
          },
          "403": {
            "description": "Forbidden - Admin access required"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/users/me/password": {
      "patch": {
        "operationId": "UsersController_changePassword",
        "summary": "Change current user password",
        "description": "Changes the password for the currently authenticated user.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Password changed successfully"
          },
          "400": {
            "description": "Current password is incorrect"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
    "/api/v1/users/{id}/permanent": {
      "delete": {
        "operationId": "UsersController_hardDelete",
        "summary": "Permanently delete user",
        "description": "Permanently deletes a user account. This action cannot be undone.",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "description": "User unique identifier",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User permanently deleted"
          },
          "404": {
            "description": "User not found"
          }
        },
        "tags": [
          "Users"
        ],
        "security": [
          {
            "JWT-auth": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "♻️ Recycling Centers Management API",
    "description": "Enterprise-grade API for managing recycling centers, materials, and user operations",
    "version": "1.0",
    "contact": {
      "name": "The Wealthiest Programmer",
      "url": "https://github.com/wealthiest-programmer",
      "email": "contact@wealthiest-programmer.dev"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Enter JWT token",
        "in": "header"
      }
    },
    "schemas": {
...
      "CreateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com"
          },
          "username": {
            "type": "string",
            "description": "Unique username (3-30 characters, lowercase letters, numbers, and underscores only)",
            "example": "marcos_enjuto"
          },
          "password": {
            "type": "string",
            "description": "User password (minimum 6 characters)",
            "example": "securePassword123",
            "minLength": 6
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "role": {
            "type": "string",
            "description": "User role in the system",
            "enum": [
              "ADMIN",
              "MANAGER",
              "USER"
            ],
            "example": "USER"
          },
          "avatar": {
            "type": "string",
            "description": "Avatar URL",
            "example": "https://example.com/avatar.jpg"
          }
        },
        "required": [
          "email",
          "username",
          "password"
        ]
      },
      "UserResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "description": "User ID"
          },
          "email": {
            "type": "string",
            "description": "User email address"
          },
          "username": {
            "type": "string",
            "description": "Username"
          },
          "firstName": {
            "type": "string",
            "description": "First name"
          },
          "lastName": {
            "type": "string",
            "description": "Last name"
          },
          "role": {
            "type": "string",
            "description": "User role",
            "enum": [
              "ADMIN",
              "MANAGER",
              "USER"
            ]
          },
          "isActive": {
            "type": "boolean",
            "description": "Account status"
          },
          "avatar": {
            "type": "string",
            "description": "Avatar URL"
          },
          "createdAt": {
            "format": "date-time",
            "type": "string",
            "description": "Account creation date"
          },
          "updatedAt": {
            "format": "date-time",
            "type": "string",
            "description": "Last update date"
          }
        },
        "required": [
          "id",
          "email",
          "username",
          "role",
          "isActive",
          "createdAt",
          "updatedAt"
        ]
      },
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "description": "User email address",
            "example": "john.doe@example.com"
          },
          "username": {
            "type": "string",
            "description": "Unique username (3-30 characters, lowercase letters, numbers, and underscores only)",
            "example": "marcos_enjuto"
          },
          "firstName": {
            "type": "string",
            "description": "User first name",
            "example": "John"
          },
          "lastName": {
            "type": "string",
            "description": "User last name",
            "example": "Doe"
          },
          "role": {
            "type": "string",
            "description": "User role in the system",
            "enum": [
              "ADMIN",
              "MANAGER",
              "USER"
            ],
            "example": "USER"
          },
          "isActive": {
            "type": "boolean",
            "description": "Account status",
            "example": true
          },
          "avatar": {
            "type": "string",
            "description": "Avatar URL",
            "example": "https://example.com/avatar.jpg"
          }
        }
      },
...
    }
  }
}